Modify the assembly_study flows for miassembler v3
I've added also the "contaminant_reference" in the assemblies metadata as this could be useful to query later.

All assemblies should be decontaminated of human data, that is done in the pipeline configuration. But't it's possible to specify a different human reference of phix reference (not likely but... you never know)

# Conflicts:
#	workflows/flows/assemble_study_tasks/make_samplesheets.py